requirements.txt
Django==3.0.8
psycopg2==2.8.5
djangorestframework==3.12.1
